update and add

+++ copy_notebook.py
@@ -4,11 +4,11 @@ from pathlib import Path
 
 notebook_files = Path("notebooks").glob("session*rdkm.ipynb")
 latest_file = max(notebook_files, key=lambda x: x.lstat().st_mtime)
-print(latest_file)
 
 filenum = int(re.search("\d+", latest_file.name).group())
 
 new_file = Path("notebooks") / f"session{filenum}_jhr.ipynb"
 
-copyfile(latest_file, new_file)
+if not new_file.exists():
+    copyfile(latest_file, new_file)
